Refactor: Object received by favourite and game
