+++ src/components/GameFilter.js
@@ -1,16 +1,16 @@
 /* eslint-disable react/no-array-index-key, react/forbid-prop-types */
 import { PropTypes } from 'prop-types';
 
-const GameFilter = ({ filteredCategories, handleFilterChange }) => (
+const GameFilter = ({ filteredGames, handleFilterChange }) => (
   <div className="text-center pb-3">
     <select
       className="form-select filter-list"
       onChange={(event) => handleFilterChange(event)}
     >
       <option value="All">All</option>
-      {filteredCategories.map((category, i) => (
-        <option key={i} value={category}>
-          {category}
+      {filteredGames.map((game, i) => (
+        <option key={i} value={game}>
+          {game}
         </option>
       ))}
     </select>
@@ -18,7 +18,7 @@ const GameFilter = ({ filteredCategories, handleFilterChange }) => (
 );
 
 GameFilter.propTypes = {
-  filteredCategories: PropTypes.array.isRequired,
+  filteredGames: PropTypes.array.isRequired,
   handleFilterChange: PropTypes.func.isRequired,
 };
